separate calendar Ui states and fix deprecations
